requirements.txt
python-telegram-bot==20.7
google-generativeai==0.8.0

mem0ai==0.1.16
chromadb==0.4.22

pydantic==2.7.4
pydantic-settings>=2.2.1,<3.0.0
numpy==1.26.4
pandas==2.0.3
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.7

# чтение PDF из ./docs и/или CV.pdf
PyPDF2==3.0.1

# Speech (Google Cloud)
google-cloud-speech==2.28.0
google-cloud-texttospeech==2.21.0

# Webhook on Vercel
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.25.2
